# Conflicts:
#	libraries/grpc-sdk/src/classes/ConduitModule.ts
#	libraries/grpc-sdk/src/index.ts
#	libraries/grpc-sdk/src/modules/config/index.ts
#	libraries/module-tools/README.mdx
#	libraries/module-tools/package.json
#	libraries/module-tools/src/ManagedModule.ts
#	libraries/module-tools/src/classes/ConduitServiceModule.ts
#	libraries/module-tools/src/utilities/index.ts
#	libraries/module-tools/src/utilities/initializeSdk.ts
#	modules/authorization/src/Authorization.ts
#	modules/database/src/database.proto
#	modules/email/src/admin/index.ts
#	modules/forms/src/Forms.ts
#	modules/functions/src/Functions.ts
#	modules/functions/src/admin/index.ts
#	modules/functions/src/controllers/utils.ts
#	packages/core/src/GrpcServer.ts

+++ modules/authentication/src/handlers/common.ts
@@ -42,10 +42,18 @@ export class CommonHandlers implements IAuthenticationStrategy {
     if (!oldRefreshToken.user) {
       throw new GrpcError(status.PERMISSION_DENIED, 'Invalid user');
     }
+    if (!(oldRefreshToken.user as User).active) {
+      throw new GrpcError(status.PERMISSION_DENIED, 'User is blocked');
+    }
 
     // delete the old refresh token
     await RefreshToken.getInstance().deleteOne({ _id: oldRefreshToken._id });
-
+    // delete all expired tokens
+    RefreshToken.getInstance()
+      .deleteMany({
+        expiresOn: { $lte: new Date() },
+      })
+      .catch();
     return TokenProvider.getInstance().provideUserTokens({
       user: oldRefreshToken.user as User,
       clientId,
